logging and failure modes
Updated the logging to be more verbose for our splunk queries later.

Added code to uncomment when we need to just skip Duo entirely
(after the Aug 20 and 29 2018 DUO1 failures).

We have this working with Refeds MFA required as well our default
mode which is to require duo only for those users who have enrolled on
our campus.